+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 - revert debian packagename from `deltachat` back to `deltachat-desktop` #4266
 - style: fix VCard color being too bright in dark theme #4255
 - remove unnecessary horizontal scrollbar in "View Group" dialog #4254
-- change title of EditAccountAndPasswordDialog to make it clearer that it is about email account #4271
+- change title and button label of EditAccountAndPasswordDialog to make it clearer that it is about email account #4271, #4279
 - fix styling of progressbars in light theme #4274
 - fix not being able to remove avatar for a mailing list #4270